Little change

+++ dbms/src/Flash/Coprocessor/TiDBTime.h
@@ -36,7 +36,7 @@ public:
     }
     MyDateTime my_date_time;
     UInt8 time_type;
-    Int8 fsp;
+    Int32 fsp;
     UInt64 toChunkTime() const
     {
         UInt64 ret = 0;
